Feat: Refactored api.py into routes. Fixes #232
